# api/openapi.yaml
openapi: 3.1.0
info:
  contact:
    name: CDE Platform engineering
  description: |
    Common Data Environment API for construction and asset information.

    ## Tenant scoping

    Every request is scoped to exactly one tenant, and the tenant is taken **from the authenticated principal** — never from a path, query, header, or body parameter. There is no tenant selector on any endpoint, and supplying one has no effect. A resource belonging to another tenant is not visible: reading it by id returns `404`, not `403`, so the API does not confirm that an id exists somewhere else.

    ## Errors

    Every error response is an RFC 9457 Problem Details document with the media type `application/problem+json`. It carries a `traceId` that identifies the request in the server logs — quote it when reporting a problem. The `type` member is a stable relative URI identifying the kind of failure, and is safe to branch on when the status code alone is too coarse; `detail` is human-readable text whose wording may change, and is not.

    Validation failures return `422` with an `invalidFields` array naming each rejected field. A request body that cannot be parsed at all returns `400`, because no field can be named.

    ## Rate limits

...

    ## Tracing

    Every response carries `X-Trace-Id`. Send a W3C `traceparent` header and its trace id is adopted, so a client-side trace and the server-side logs share one identifier.
  license:
    name: Proprietary
  title: CDE Platform API
...
servers:
...
security:
- bearerAuth: []
...
paths:
...
  /api/documents/{documentId}/versions:
    get:
      description: |-
        Every version the document has had, newest first, each with what produced it and who committed it.

        A document nothing has processed yet still has a history: its upload is recorded as version 1 on first read, so the original is downloadable whether or not anyone has opened the history panel.

        Requires the `document:read` permission.
      operationId: listDocumentVersions
      parameters:
      - description: Identifier of the document.
        example: 1180
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      responses:
        "200":
          content:
            application/json:
              schema:
                type:
                - array
                example: null
                items:
                  $ref: "#/components/schemas/DocumentVersionResponse"
                  example: null
          description: "The version history, newest first."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: No document with that id is visible to the caller.
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Read a document's version history
      tags:
      - Document versions
  /api/documents/{documentId}/versions/{versionNumber}/file:
    get:
      description: |-
        Returns the bytes as they stood at that version, not the current ones. This is how a signature on an earlier version stays checkable after the document has moved on.

        The reply is a file download, not JSON.

        Requires the `document:read` permission.
      operationId: downloadDocumentVersion
      parameters:
      - description: Identifier of the document.
        example: 1180
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      - description: "Version to download, counting from 1."
        example: 2
        in: path
        name: versionNumber
        required: true
        schema:
          type:
          - integer
          format: int32
          example: null
      responses:
        "200":
          content:
            application/pdf:
              schema:
                format: binary
                example: null
          description: The version's bytes.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No such document, no such version, or its file is no longer\
            \ on disk."
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Download one version of a document
      tags:
      - Document versions
  /api/documents/{documentId}/versions/{versionNumber}/restore:
    post:
      description: |-
        Copies the chosen version forward as a new one rather than discarding the versions after it. Nothing that references a later version — a signature above all — is left pointing at bytes that no longer exist.

        The reply describes the version this created, not the one that was restored from.

        Requires the `document:write` permission.
      operationId: restoreDocumentVersion
      parameters:
      - description: Identifier of the document.
        example: 1180
        in: path
        name: documentId
        required: true
        schema:
          type:
          - integer
          format: int64
          example: null
      - description: "Version to restore, counting from 1."
        example: 2
        in: path
        name: versionNumber
        required: true
        schema:
          type:
          - integer
          format: int32
          example: null
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentVersionResponse"
                example: null
          description: "The new version, holding the restored content."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "The request could not be read — malformed JSON, or a value\
            \ outside an accepted set. No field can be named, because parsing did\
            \ not get that far."
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No credential was presented, or the one presented is expired\
            \ or invalid."
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "Authenticated, but without the permission this operation requires."
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "No such document, or no such version of it."
        "429":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: "A rate limit was exceeded. `Retry-After` gives the seconds\
            \ to wait. Limits apply per account, per source address and per tenant\
            \ independently."
        "500":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
                example: null
          description: The request failed for a reason the caller cannot act on. Quote
            the `traceId` when reporting it.
      summary: Make an earlier version current again
      tags:
      - Document versions
...
components:
  schemas:
...
    DocumentVersionResponse:
      description: "One entry in a document's history. Every operation that rewrites\
        \ a document commits one of these, and every earlier entry stays downloadable."
      example: null
      properties:
        contentHash:
          type:
          - string
          description: "SHA-256 of this version's bytes, so a downloaded copy can\
            \ be checked against what the server holds."
          example: 9f2c4b7e1a8d3f60c5b2e9a47d18f3c06b5e2a91d7f4c308b6a1e5d9c2f70348
          pattern: "^[0-9a-f]{64}$"
        createdAt:
          type:
          - string
          format: date-time
          description: "When this version was committed, UTC."
          example: 2026-02-21T10:33:09
          readOnly: true
        createdBy:
          type:
          - string
          description: Sign-in name of the account that committed it.
          example: j.okafor
          readOnly: true
        current:
          type:
          - boolean
          description: Whether this is the version the document currently serves.
          example: false
        fileName:
          type:
          - string
          description: File name this version was written under.
          example: RVD-XX-02-DR-A-1200.pdf
        fileSize:
          type:
          - integer
          format: int64
          description: Size of this version in bytes.
          example: 4721664
          minimum: 0
        operation:
          type:
          - string
          description: What produced this version.
          enum:
          - UPLOAD
          - REDACT
          - OCR
          - FLATTEN
          - FORM_FILL
          - FORM_DESIGN
          - PAGES
          - SIGN
          - RESTORE
          example: REDACT
        summary:
          type:
          - string
          description: "What changed, in one line, written for the history panel."
          example: Redacted 6 regions across 2 pages.
        version:
          type:
          - integer
          format: int32
          description: "Version number, counting from 1 for the original upload."
          example: 3
          minimum: 1
      required:
      - createdAt
      - current
      - operation
      - summary
      - version
...
    ProblemDetail:
      type:
      - object
      description: "RFC 9457 Problem Details. The single error shape returned by every\
        \ endpoint, for every status."
      example: null
      properties:
        detail:
          type:
          - string
          description: "What happened and what to do next, written for a person to\
            \ read. Never contains a stack trace, SQL, or an internal type name."
          example: 1 field is invalid.
        instance:
          type:
          - string
          format: uri-reference
          description: Path of the request that failed.
          example: /api/projects
        invalidFields:
          type:
          - array
          description: "Present on `422` validation failures. Names every field that\
            \ was rejected, so a client can attach each message to the control it\
            \ belongs to rather than to the form as a whole."
          example: null
          items:
            type:
            - object
            description: One field the request was rejected for.
            example: null
            properties:
              field:
                type:
                - string
                description: "Path of the rejected field, as it appears in the request\
                  \ body."
                example: name
              message:
                type:
                - string
                description: "What is wrong with the value, phrased for the person\
                  \ who typed it."
                example: must not be blank
            required:
            - field
            - message
        status:
          type:
          - integer
          format: int32
          description: "HTTP status code, repeated here so a logged body is self-contained."
          example: 422
          maximum: 599
          minimum: 400
        timestamp:
          type:
          - string
          format: date-time
          description: "When the failure was recorded, UTC."
          example: 2026-02-17T09:41:12.884Z
        title:
          type:
          - string
          description: Short summary of the problem type. Does not vary between occurrences.
          example: Validation failed
        traceId:
          type:
          - string
          description: "Identifies this request in the server logs, and is also returned\
            \ in the `X-Trace-Id` header. Quote it when reporting a problem."
          example: 4f8a1c2e9b7d6a5f3e2d1c0b9a8f7e6d
          pattern: "^[0-9a-f]{32}$"
        type:
          type:
          - string
          format: uri-reference
          description: "Stable identifier for this kind of problem, relative to the\
            \ API host. Safe to branch on — unlike `detail`, its value does not change\
            \ with wording."
          example: /problems/validation-failed
      required:
      - type
      - title
      - status
      - detail
      - traceId
...
  securitySchemes:
    apiKey:
      description: "Machine credential, sent as `X-API-Key`. Shown once at creation\
        \ and stored only as a SHA-256 digest, so a lost key is replaced rather than\
        \ recovered. Scoped, expiring, and rate-limited per key."
      in: header
      name: X-API-Key
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      description: "Signed JSON Web Token from `POST /api/auth/login`, sent as `Authorization:\
        \ Bearer <token>`. The token carries the subject and the tenant; the tenant\
        \ claim is authoritative and is not overridable by any request parameter."
      scheme: bearer
      type: http
    oauth2:
      description: Authorization code with PKCE for user-facing clients; client credentials
        for machine clients. The implicit flow is not offered.
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          refreshUrl: /oauth2/token
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
            project:write: "Create, update and delete projects"
            signature:read: Read and verify digital signatures
            signature:write: Apply a digital signature to a document version
          tokenUrl: /oauth2/token
        clientCredentials:
          scopes:
            annotation:read: Read markup on a document
            annotation:write: "Create, update and resolve markup"
            document:process: Run operations that rewrite a document
            document:read: Read document metadata and content
            document:write: "Upload, replace and delete documents"
            project:read: List and read projects
...